Handle length fields correctly in LCM marshaling functions
This concerns the array length fields (usually "num_*") present in the
ASPN YAMLs, aspn-c, aspn-lcm, and aspn-lcm-py but intentionally left out
of aspn-py and aspn-cpp. The py-to-lcm marshaling functions must create
these fields based on the length of the corresponding arrays. This is
currently done via hardcoded special case logic, and it has two issues:

1. The creation of these fields in the marshaling functions is currently
   bidirectional, so they are accidentally created when converting from
   lcm-to-py. This is undesirable, since the fields are not declared in
   the aspn-py class files.

2. The following fields are not being created in aspn-lcm-py, despite
   being declared in the aspn-lcm-py class files:
   - num_integrity (with the integrity field, in many files)
   - num_meas (with the covariance field, in many files)
   - num_obs (with the obs field in measurement_direction_3d_to_points
     and measurement_direction_2d_to_points)
   - num_position_components (with the position_covariance field in
     type_remote_point)

Replace the hardcoded logic with a simple, dynamic approach which sets
all length fields appropriately, and only when converting py-to-lcm.

Closes #64