Simplify aspn_yaml_to_lcm_translations.py assignments
Non-functional change.

Previously, the assignments list was split into assignments
(bidirectional), assignments_to_lcm, and assignments_from_lcm. This
results in needless complication, and it has the side effect of changing
the order of the assignments in the marshaling function.

Instead, just query self.current_struct.to_lcm to determine what to add
to assignments.