Properly handle optional (nullable) fields in LCM translations
Currently, if a optional ASPN field is set to None in aspn-py, the
py-to-lcm marshaling functions will copy this None value. This will
break the LCM transport since it cannot handle None fields. Instead, the
marshaling functions should default-initialize such fields when
converting to aspn-lcm-py.

The inverse problem--converting default-initialized fields to None in
aspn-py--is difficult to fix, since we don't know if these fields should
be None or present with the default value. To be safe, let's leave them
as is for now.

Closes #65