Generate LCM marshaling utility maps
In practice, it is helpful to have maps which aid in finding the desired
marshaling functions (based on the type of the input object). Currently,
such maps are being constructed by users (e.g. in pntos-python), but it
makes sense to auto-generate them and include them with the marshaling
functions for ease of use in any project.

In addition, when decoding LCM, it is helpful to have a map from the
fingerprint of each LCM Python type to the corresponding decode()
method. Auto-generate this as well.