Generate aspn-ros <---> aspn-ros-py Python marshaling functions
To help Python ROS nodes convert between aspn-ros and aspn-ros-py, we
generate marshaling functions for each ASPN type, measurement, and
metadata. These take care of the differences in how enums, arrays,
matrices, etc. are handled between Python ROS and aspn-py.

Note that fields cannot be None in ROS, so optional fields missing in
aspn-py are default-initialized when converting to aspn-ros-py.

Note that the length fields ("num_*", e.g.) currently missing in aspn-py
but present in aspn-ros are generated by the marshaling functions when
converting to ROS.

Also stage the metadata files needed to put these in a ROS package.

See #66