Make AspnRosNode queue size configurable
The queue size (QoS parameter) determines how many messages are buffered
by the ROS messaging system when publishing/subscribing. 10 is a
reasonable default, but when replaying bagfiles at high speeds, it
proves insufficient and leads to dropped messages. For this reason,
default to 100 instead, but also allow the user to specify a custom
value upon node instantiation.