ROS Docker containers: run clang-format on output of colcon build
Immediate motivation: colcon build (run by the ROS Ubuntu development
Docker containers) generates __description.c files for each message,
each of which has a section labeled by a comment as "// Include
directives for referenced types". But with ROS Jazzy, the exact order of
the #includes seem to vary from build to build (perhaps due to a race
condition), causing instability. This will lead to needless changes in
firehose-outputs.

Also, it's probably a good idea in general to format the C(++) files
generated by ROS in case any unfortunate human has to read them.

So install clang-format in the ROS Docker containers and run it on all
.c/.h/.cpp/.hpp files generated by the colcon build.