aspn-py: type-hint numpy arrays better
We are about to add mypy checking of ASPN-PY to the CI. First, we must
fix the existing mypy errors. All of these are caused by how Numpy
arrays are handled in the generated code.

The backwards-compatible support for using np.ndarray as a type-hint was
causing some of the errors. It is simplest to just use NDArray,
introduced in Numpy v1.21. As this is a reasonably old version, just
update the staged ASPN-PY pyproject.toml to specify numpy>=1.21 as a
dependency.

In addition, the use of custom type aliases (NumpyArray and NumpyMatrix)
for 1D vs 2D arrays is not helpful, since Numpy doesn't have a good to
specify 1D vs 2D dimensionality to type checkers (newer versions support
specifying precise array dimensions). In fact, both aliases were set to
the same thing (either np.ndarray or NDArray). I've found it difficult
to satisfy both mypy and my LSP, pyright, while using these aliases.
Instead, just type-hint arrays a NDArrays.

Finally, in order to satisfy mypy and pyright, the dtype specification
on the type-hint of an array should be a Numpy dtype (e.g. np.float64),
not a Python datatype (e.g. float). Update the ASPN-PY code gen to
programmatically generate the correct dtype in a NDArray type-hint.

With these, changes mypy --strict yields no more errors in ASPN-PY.

See #76